+++ mz_strm.h
@@ -99,7 +99,7 @@ void    mz_stream_delete(void **stream);
 
 /***************************************************************************/
 
-typedef int32_t (*mz_stream_crc32_update)(int32_t value, const void *buf, int32_t size);
+typedef int64_t (*mz_stream_crc32_update)(int64_t value, const void *buf, int32_t size);
 
 int32_t mz_stream_crc32_open(void *stream, const char *filename, int32_t mode);
 int32_t mz_stream_crc32_is_open(void *stream);
